phoenixd: add TCP channel
Adds the ability to communicate using TCP with the `socat`
tool or the `OpenOCD` debugger providing an RTT or
semihosting channel.

The data is transferred in a form such as via a serial
port or via a USB-uart adapter, as understood by the uart
driver in `plo`, the TCP channel acts as a tunnel here.

JIRA: RTOS-532